requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
appnope=0.1.4=pyhd8ed1ab_1
asttokens=3.0.0=pyhd8ed1ab_1
blinker=1.9.0=pypi_0
bzip2=1.0.8=h80987f9_6
ca-certificates=2024.11.26=hca03da5_0
cachetools=5.5.0=pypi_0
certifi=2024.8.30=pypi_0
charset-normalizer=3.4.0=pypi_0
click=8.1.7=pypi_0
comm=0.2.2=pyhd8ed1ab_1
debugpy=1.6.7=py313h313beb8_0
decorator=5.1.1=pyhd8ed1ab_1
dnspython=2.7.0=pypi_0
email-validator=2.2.0=pypi_0
exceptiongroup=1.2.2=pyhd8ed1ab_1
executing=2.1.0=pyhd8ed1ab_1
expat=2.6.3=h313beb8_0
flask=3.1.0=pypi_0
flask-mongoengine=1.0.0=pypi_0
flask-pymongo=2.3.0=pypi_0
flask-wtf=1.2.2=pypi_0
google-auth=2.37.0=pypi_0
google-auth-oauthlib=1.2.1=pypi_0
gspread=6.1.4=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.5.0=pyha770c72_1
ipykernel=6.29.5=pyh57ce528_0
ipython=8.30.0=pyh707e725_0
itsdangerous=2.2.0=pypi_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.4=pypi_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.7.2=pyh31011fe_1
libcxx=14.0.6=h848a8c0_0
libffi=3.4.4=hca03da5_1
libmpdec=4.0.0=h80987f9_0
libsodium=1.0.18=h27ca646_1
load-dotenv=0.1.0=pypi_0
markupsafe=3.0.2=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mongoengine=0.29.1=pypi_0
ncurses=6.4=h313beb8_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
numpy=2.1.3=pypi_0
oauthlib=3.2.2=pypi_0
openssl=3.4.0=h39f12f2_0
packaging=24.2=pyhd8ed1ab_2
pandas=2.2.3=pypi_0
parso=0.8.4=pyhd8ed1ab_1
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pip=24.2=py313hca03da5_0
platformdirs=4.3.6=pyhd8ed1ab_1
prompt-toolkit=3.0.48=pyha770c72_1
psutil=6.1.0=py313h63a2874_0
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pygments=2.18.0=pyhd8ed1ab_1
pymongo=4.10.1=pypi_0
python=3.13.0=h4862095_100_cp313
python-dateutil=2.9.0.post0=pyhff2d567_1
python-dotenv=1.0.1=pypi_0
python_abi=3.13=0_cp313
pytz=2024.2=pypi_0
pyzmq=25.1.2=py313h313beb8_0
readline=8.2=h1a28f6b_0
requests=2.32.3=pypi_0
requests-oauthlib=2.0.0=pypi_0
rsa=4.9=pypi_0
setuptools=75.1.0=py313hca03da5_0
six=1.17.0=pyhd8ed1ab_0
sqlite=3.45.3=h80987f9_0
stack_data=0.6.3=pyhd8ed1ab_1
tk=8.6.14=h6ba3021_0
tornado=6.4.2=py313h90d716c_0
traitlets=5.14.3=pyhd8ed1ab_1
typing_extensions=4.12.2=pyha770c72_1
tzdata=2024.2=pypi_0
urllib3=2.2.3=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_1
werkzeug=3.1.3=pypi_0
wheel=0.44.0=py313hca03da5_0
wtforms=3.2.1=pypi_0
xz=5.4.6=h80987f9_1
zeromq=4.3.5=h313beb8_0
zipp=3.21.0=pyhd8ed1ab_1
zlib=1.2.13=h18a0788_1
